gcp-cloud-infra/terraform/modules/virus-scan/requirements.txt
# requirements.txt
google-cloud-storage
google-cloud-firestore
